+++ client/src/components/signup/SignupBirth/SignupBirth.jsx
@@ -1,8 +1,7 @@
-import React, {useState} from 'react';
-import { BirthMonth, BirthYear } from '../../styles/common/input';
-import { ValidCheck } from '../../styles/common/validtext';
-import { InputBox, BirthBox } from '../../styles/signup/container';
-import { SignupTitle } from '../../styles/signup/title';
+import React, { useState } from 'react';
+import SignupBirthStyle from './SignupBirth.style';
+
+// 오류 메세지 출력하는 과정에서 문제가 있음
 
 function SignupBirth ({inputs, setInputs}) {
   const [message, setMessage] = useState({
@@ -58,31 +57,12 @@ function SignupBirth ({inputs, setInputs}) {
   }
 
   return (
-    <InputBox>
-      <SignupTitle> 생년월일 </SignupTitle>
-      <BirthBox>
-        <BirthYear type="text" name="yy" onBlur={handleBirthYearInput} placeholder='년(4자)' maxLength={4}/>
-        <BirthMonth name="mm" onBlur={handleBirthMonthInput}>
-          <option value="00"> 월 </option>
-          <option value="01"> 1 </option>
-          <option value="02"> 2 </option>
-          <option value="03"> 3 </option>
-          <option value="04"> 4 </option>
-          <option value="05"> 5 </option>
-          <option value="06"> 6 </option>
-          <option value="07"> 7 </option>
-          <option value="08"> 8 </option>
-          <option value="09"> 9 </option>
-          <option value="10"> 10 </option>
-          <option value="11"> 11 </option>
-          <option value="12"> 12 </option>
-        </BirthMonth>
-        <BirthYear type="text" name="dd" onBlur={handleBirthDayInput} placeholder='일' maxLength={2}/>
-      </BirthBox>
-      {message.yy !== "" && <ValidCheck> {message.yy} </ValidCheck>}
-      {message.mm !== "" && <ValidCheck> {message.mm} </ValidCheck>}
-      {message.dd !== "" && <ValidCheck> {message.dd} </ValidCheck>}
-    </InputBox>
+    <SignupBirthStyle
+      message={message}
+      onBirthYearInput={handleBirthYearInput}
+      onBirthMonthInput={handleBirthMonthInput}
+      onBirthDayInput={handleBirthDayInput}
+    />
   )
 }
