+++ client/src/components/signup/SignupButton/SignupBtns.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignupBtnsStyle from './SignupBtns.style';
+
+function SignupBtns ({ onSignupSubmit }) {
+  return (
+    <SignupBtnsStyle 
+      onSignupSubmit={onSignupSubmit}
+    />
+  );
+}
+
+export default SignupBtns;
