+++ client/src/pages/SignupPage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SingupSection from "../components/signup/SignupSection";
+import SingupSection from "../components/signup/SignupSection/SignupSection";
 
 export default function SignupPage() {
   return <SingupSection />;
